requirement.txt
absl-py==1.3.0
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
astor==0.8.1
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
click==8.1.3
cmake==3.25.0
colorama==0.4.6
cycler==0.11.0
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e2vdu8snoo/croots/recipe/django_1661284240723/work
django-cors-headers==3.13.0
djongo==1.3.6
dlib @ file:///C:/Users/ojyse/PycharmProjects/pythonProject/Face-Clustering-using-DBSCAN/dlib-19.17.0-cp37-cp37m-win_amd64.whl
dnspython==2.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fanstatic==1.2
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.50.0
h5py==2.10.0
imageio==2.22.4
importlib-metadata==5.0.0
joblib==1.1.0
js==1.0
Keras==2.1.6
Keras-Applications==1.0.8
keras-facenet==0.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mglearn==0.1.9
mtcnn==0.1.1
numpy==1.18.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
protobuf==3.20.3
psycopg2 @ file:///C:/b/abs_6cdgqr8yu7/croot/psycopg2_1666817457048/work
PyJWT==2.6.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
shutilwhich==1.1.0
six==1.16.0
sklearn==0.0.post1
sqlparse==0.2.4
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-directml==1.15.8
tensorflow-estimator==1.15.1
termcolor==2.1.0
threadpoolctl==3.1.0
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
WebOb==1.8.7
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.10.0
